requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.4.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
gast==0.3.3
google-api-core==1.21.0
google-auth==1.18.0
google-cloud-core==1.3.0
google-cloud-storage==1.29.0
google-cloud-vision==1.0.0
google-resumable-media==0.5.1
googleapis-common-protos==1.52.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
mock==4.0.2
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.1.2
protobuf==3.12.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
scipy==1.5.1
six==1.15.0
sqlparse==0.3.1
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
urllib3==1.25.9
Werkzeug==1.0.1
zipp==3.1.0
